Componentized form as TodoForm
Refactored: extracted form from App.vue into TodoForm component
