+++ src/types.ts
@@ -1,0 +1,5 @@
+export interface Task {
+  id: string;
+  title: string;
+  done: boolean;
+}
